simplify architecture - i way overdid it
