+++ ui/packages/comhairle/src/lib/tools/prioritization/PrioritizationReport.svelte
@@ -1,7 +1,3 @@
-<script lang="ts">
-	/* Keeping a typed stub so the mode switch in PrioritizationTool still resolves. */
-</script>
-
 <section class="prioritization-report">
 	<p><em>Report mode coming soon — needs backend support.</em></p>
 </section>
